Moves attack builders in more specific packages
