+++ src/test/scala/it/unibo/authsim/library/dsl/attack/builders/DictionaryAttackBuilderTest.scala
@@ -1,5 +1,6 @@
 package it.unibo.authsim.library.dsl.attack.builders
 
+import it.unibo.authsim.library.dsl.attack.builders.offline.bruteforce.DictionaryAttackBuilder
 import it.unibo.authsim.library.dsl.attack.statistics.Statistics
 import it.unibo.authsim.library.dsl.consumers.StatisticsConsumer
 import it.unibo.authsim.library.dsl.policy.model.Policy
